[Loom] Preserve ownership ties in assembly result annotations
Low packet ownership ties can differ from an instruction descriptor's
physical register constraints. Requiring their tie sets to match
prevented valid operations from printing as assembly.

Use the canonical result annotation grammar for mnemonic packets. An
explicit result list supplies the complete ownership tie set, with plain
types expressing independent results and operand references expressing
consumed values. Omitted annotations retain descriptor inference, and
printing adds annotations whenever inference would change ownership.

Share result parsing and printing with canonical operations and keep the
tied-result printer counter wide enough for the IR count domain. Author
the affected native load and spill cases in assembly, with source
roundtrip coverage in Loom fixtures.

+++ loom/src/loom/format/text/printer/format_signatures.c
@@ -91,7 +91,7 @@ iree_status_t loom_print_result_type_list(
 
     const loom_tied_result_t* tied = loom_op_tied_results(op);
     bool is_tied = false;
-    for (uint8_t t = 0; t < op->tied_result_count; ++t) {
+    for (uint16_t t = 0; t < op->tied_result_count; ++t) {
       if (tied[t].result_index == j) {
         if (tied[t].operand_index >= tied_operand_count || !tied_operand_ids) {
           return iree_make_status(
